refactor: auth에서 authToken으로 이름변경
