--- mentors-domain/src/main/java/com/mentors/auth/domain/Auth.java
@@ -1,9 +1,0 @@
-package com.mentors.auth.domain;
-
-import java.time.LocalDateTime;
-
-public record Auth(Long memberId,
-                   String refreshToken,
-                   LocalDateTime createdAt,
-                   LocalDateTime updatedAt) {
-}
