+++ mentors-domain/src/main/java/com/mentors/authToken/domain/AuthToken.java
@@ -1,0 +1,9 @@
+package com.mentors.authToken.domain;
+
+import java.time.LocalDateTime;
+
+public record AuthToken(Long memberId,
+                        String refreshToken,
+                        LocalDateTime createdAt,
+                        LocalDateTime updatedAt) {
+}
