--- mentors-infrastruct/persistence-database/src/main/java/com/mentors/auth/AuthRepository.java
@@ -1,6 +1,0 @@
-package com.mentors.auth;
-
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface AuthRepository extends JpaRepository<AuthEntity, Long> {
-}
