+++ mentors-infrastruct/persistence-database/src/main/java/com/mentors/authToken/AuthTokenRepository.java
@@ -1,0 +1,6 @@
+package com.mentors.authToken;
+
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface AuthTokenRepository extends JpaRepository<AuthTokenEntity, Long> {
+}
